(Update) deeptech.id Implementing Template in .src
- Combine gatsby-config.js (root) with gatsby-config.js Novela file
- Move all file and folder in .src Novela to .src root

Obstacle (error.log)
A gatsby-config exported as a Function can only be used as a theme and not run directly.
If you are trying to run a theme directly, use the theme in an example site or starter instead and run that site to test.
If you are in the root gatsby-config.js for your site, change the export to be an object and not a function as functionsare not supported in the root gatsby-config.

Co-Authored-By: Zain Fathoni <zainfathoni@users.noreply.github.com>